# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Smart Airport Ride Pooling System API
  version: 1.0.0
  description: A comprehensive API for airport ride pooling with real-time matching and dynamic pricing
  contact:
    name: API Support
    email: support@airportpooling.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
paths:
...
  /api/rides/request:
    post:
      tags:
        - Rides
      summary: Create a new ride request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RideRequest'
      responses:
        '201':
          description: Ride request created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      ride:
                        $ref: '#/components/schemas/Ride'
                      pricing:
                        type: object
        '401':
          description: Authentication required
        '429':
          description: Too many ride requests
  /api/rides/estimate:
    post:
      tags:
        - Rides
      summary: Get price estimate without creating a ride
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pickup
                - destination
                - passengers
                - luggage
              properties:
                pickup:
                  type: object
                  required:
                    - address
                    - coordinates
                  properties:
                    address:
                      type: string
                    coordinates:
                      type: array
                      items:
                        type: number
                      minItems: 2
                      maxItems: 2
                destination:
                  type: object
                  required:
                    - airportCode
                    - name
                    - coordinates
                  properties:
                    airportCode:
                      type: string
                      enum: [BOM, DEL, BLR, HYD, MAA]
                    name:
                      type: string
                    coordinates:
                      type: array
                      items:
                        type: number
                      minItems: 2
                      maxItems: 2
                passengers:
                  type: integer
                  minimum: 1
                  maximum: 6
                luggage:
                  type: integer
                  minimum: 0
                  maximum: 5
                isPool:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Price estimate calculated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      estimate:
                        type: object
                      surge:
                        type: object
        '429':
          description: Too many estimate requests
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Ride:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        pickup:
          type: object
          properties:
            address:
              type: string
            coordinates:
              type: array
              items:
                type: number
        destination:
          type: object
          properties:
            airportCode:
              type: string
            name:
              type: string
            coordinates:
              type: array
              items:
                type: number
        flightNumber:
          type: string
        scheduledTime:
          type: string
          format: date-time
        passengers:
          type: integer
          minimum: 1
          maximum: 6
        luggage:
          type: integer
          minimum: 0
          maximum: 5
        status:
          type: string
          enum: [requested, matched, assigned, in_progress, completed, cancelled]
        pricing:
          type: object
          properties:
            baseFare:
              type: number
            surgeMultiplier:
              type: number
            luggageFee:
              type: number
            poolDiscount:
              type: number
            total:
              type: number
        isPool:
          type: boolean
        createdAt:
          type: string
          format: date-time
    RideRequest:
      type: object
      required:
        - pickup
        - destination
        - flightNumber
        - scheduledTime
        - passengers
        - luggage
      properties:
        pickup:
          type: object
          required:
            - address
            - coordinates
          properties:
            address:
              type: string
              maxLength: 200
            coordinates:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
        destination:
          type: object
          required:
            - airportCode
            - name
            - coordinates
          properties:
            airportCode:
              type: string
              enum: [BOM, DEL, BLR, HYD, MAA]
            name:
              type: string
            coordinates:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
        flightNumber:
          type: string
          pattern: '^[A-Z]{2}\d{3,4}$'
        scheduledTime:
          type: string
          format: date-time
        passengers:
          type: integer
          minimum: 1
          maximum: 6
        luggage:
          type: integer
          minimum: 0
          maximum: 5
        isPool:
          type: boolean
          default: true
        detourTolerance:
          type: integer
          minimum: 0
          maximum: 60
          default: 15
        specialRequests:
          type: string
          maxLength: 500
...